trait for deal exception response

+++ app/Services/Auth/Register/RegisterService.php
@@ -2,16 +2,23 @@
 namespace App\Services\Auth\Register;
 
 use App\Models\User;
+use Illuminate\Http\Response;
+use App\Traits\ApiResponseTrait;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 use App\Events\VereficationCodeEvent;
-use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Session;
-use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Storage;
+use App\Traits\Files\FileOperationsTrait;
+use App\Exceptions\Otp\InvalidCodeException;
+use App\Exceptions\userNotFound\UserNotFoundException;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 Class RegisterService
 {
+    use FileOperationsTrait ,ApiResponseTrait;
+
     public function signUp($request)
     {
         DB::beginTransaction();
@@ -20,12 +27,18 @@ Class RegisterService
 
             if($request->hasFile('profile_photo'))
             {
-                $imagePath = $this->UploadFile('image',$request->file('profile_photo') , $requestData['user_name']);
+                $file =$request->file('profile_photo') ;
+                $imagePath = $this->generatePath('image',$file, $requestData['user_name']);
+                $this->uploadFile($file, $imagePath);
+
             }
 
             if($request->hasFile('certificate'))
             {
-                $filePath = $this->UploadFile('pdf',$request->file('certificate') , $requestData['user_name']);
+                $file =$request->file('certificate') ;
+                $filePath = $this->generatePath('pdf',$file, $requestData['user_name']);
+                $this->uploadFile($file, $filePath);
+
             }
 
             $user = User::create([
@@ -49,40 +62,45 @@ Class RegisterService
             return ['token'=> $accessToken];
 
 
-        }catch(\Exception $e)
+        }catch(Throwable $exception)
         {
             DB::rollback();
-            return $e->getMessage();
+            return $this->handleException($exception);
+
+            // return $e->getMessage();
         }
 
 
 
     }
 /************************************************************************************/
-    public function UploadFile($file_type,$image , $user_name)
+    public function generatePath($file_type,$image , $user_name)
     {
         if($file_type=='image')
         {
             $folderName = 'profile';
             $folderPath = 'images/' . $folderName . '/users/' . $user_name ;
-            $path = Storage::disk('public')->putFile($folderPath, $image);
+            // $path = Storage::disk('public')->putFile($folderPath, $image);
 
         }
         else if($file_type=='pdf')
         {
             $folderName = 'pdf';
             $folderPath = 'Files/' . $folderName . '/users/' . $user_name ;
-            $path = Storage::disk('public')->putFile($folderPath, $image);
+            // $path = Storage::disk('public')->putFile($folderPath, $image);
 
         }
 
-        return $path;
+        return $folderPath;
     }
-
     /************************************************************************************/
     public function resendVerifyCode()
     {
         $user = auth()->user();
+        if(!$user)
+        {
+            throw new AuthenticationException();
+        }
         $user->generateVerificationCode();
         return $user->verify_code;
 
@@ -91,12 +109,17 @@ Class RegisterService
     public function confirmVerifyCode($request)
     {
         $user = Auth::user();
+        if(!$user)
+        {
+            throw new AuthenticationException();
+        }
+
         $code = $request->verify_code;
         $expired_at = $user->created_at->addMinutes(3);
 
         if ($code !== $user->verify_code || now() > $expired_at)
         {
-            return false;
+            throw new InvalidCodeException();
         }else{
             $user->resetVerificationCode();
             $user->email_verified_at = now();
@@ -110,6 +133,10 @@ Class RegisterService
     public function refreshToken()
     {
         $user = Auth::user();
+        if(!$user)
+        {
+            throw new AuthenticationException();
+        }
         // $oldToken = $user->currentAccessToken()->delete();
         $user->tokens()->delete();
         $refreshToken = $user->createToken('#$_refresh_token_@#',
